Create JobLoggerStateIndicator component
